Add links container handling and check link functionality
